added porfolioList component to seperate data from the map to better update and seperate in the future
